refactor: align with naming conventions
