+++ hrm-domain/hrm-core/routes.ts
@@ -23,7 +23,7 @@ import postSurveys from './post-survey/post-survey-routes-v0';
 import referrals from './referral/referral-routes-v0';
 import permissions from './permissions/permissions-routes-v0';
 import profiles from './profile/profileRoutesV0';
-import adminProfiles from './profile/admin-profile-routes-v0';
+import adminProfiles from './profile/adminProfileRoutesV0';
 import { Permissions } from './permissions';
 
 export const HRM_ROUTES: {
